Story7

+++ Controllers/OrderController.js
@@ -1,10 +1,16 @@
 const DbService = require('../Services/DbService')
 const OrderService = require('../Services/OrderService')
+const DishesService = require('../Services/DishesService')
 const JSONResponseService = require('../Services/JSONResponseService')
+const PriceService = require('../Services/PriceService')
 const ObjectId = require('mongodb').ObjectId
+
 const orderValidate = require('../Validators/newOrderValidator.json')
 const addToOrderValidate = require('../Validators/addItemsToOrderValidator.json')
+const idRemoveValidator = require('../Validators/removeOrderItem.json')
+const submitOrderValidate = require('../Validators/submitFinalOrder.json')
 const editOrderItemQuantityValidate = require('../Validators/editOrderItemQuantityValidator.json')
+
 const Ajv = require('ajv')
 const addFormats = require('ajv-formats')
 
@@ -18,6 +24,8 @@ const createNewOrder = (req, res) => {
             firstLineOfAddress: req.body.firstLineOfAddress,
             postcode: req.body.postcode,
             email: req.body.email,
+            totalCost: 0,
+            isOrderSubmitted: false,
             orderItems: []
         }
 
@@ -72,6 +80,48 @@ const addToOrder = (req, res) => {
     })
 }
 
+
+let getDishPriceById = (req, res) => {
+    DbService.connectToDb(async (db) => {
+        const dishId = ObjectId(req.params.id)
+        const dish = await DishesService.getDishPriceById(db, dishId)
+        return res.json(dish)
+    })
+}
+
+let submitFinalOrder = (req, res) => {
+    DbService.connectToDb(async (db) => {
+        const order = {
+            orderId: ObjectId(req.body.orderId)
+        }
+
+        const compile = ajv.compile(submitOrderValidate)
+        const valid = compile(order)
+
+        if (valid) {
+            try {
+                const finalisedOrder = await OrderService.getOrderDetails(db, order.orderId)
+                const totalPrice = await PriceService.calculateTotalPrice(db, finalisedOrder)
+                let isSubmitted = finalisedOrder.isOrderSubmitted
+                if (isSubmitted !== true) {
+                    await OrderService.submitFinalOrder(db, order, totalPrice)
+                    let response = JSONResponseService.generateSuccessResponse()
+                    response.data = await OrderService.getOrderDetails(db, order.orderId)
+                    response.message = "The order has been placed"
+                    return res.json(response)
+                }
+                let response = JSONResponseService.generateFailureResponse()
+                response.message = "Order has already been submitted"
+                return res.json(response)
+            } catch (e) {
+                let response = JSONResponseService.generateFailureResponse()
+                response.message = "Database request failed"
+            }
+        }
+    })
+}
+
+
 const editOrderItemQuantity = (req, res) => {
     DbService.connectToDb(async (db) => {
         const itemsToAmend = {
@@ -107,9 +157,29 @@ const editOrderItemQuantity = (req, res) => {
     })
 }
 
-module.exports.createNewOrder = createNewOrder
-module.exports.addToOrder = addToOrder
-module.exports.editOrderItemQuantity = editOrderItemQuantity
+
+let getOrderDetails = (req, res) => {
+    DbService.connectToDb(async (db) => {
+        const order = {
+            orderId: ObjectId(req.params.id)
+        }
+        const compile = ajv.compile(submitOrderValidate)
+        const valid = compile(order)
+        if (valid) {
+            try {
+                const orderDetails = await OrderService.getOrderDetails(db, order.orderId)
+                return res.json(orderDetails)
+            } catch (e) {
+                let response = JSONResponseService.generateFailureResponse()
+                response.message = "Order number not valid"
+                return res.json(response)
+            }
+            let response = JSONResponseService.generateFailureResponse()
+            response.message = "Database request failed"
+            return res.json(response)
+        }
+    })
+}
 
 let removeOrderItem = (req, res) => {
     DbService.connectToDb(async (db) => {
@@ -117,19 +187,34 @@ let removeOrderItem = (req, res) => {
             orderId: ObjectId(req.body.orderId),
             menuItemId: req.body.menuItemId,
         }
-        const removeOrder = await OrderService.removeOrderItem(db, item)
-        if(removeOrder.modifiedCount === 1) {
-            let response  = JSONResponseService.generateSuccessResponse()
-            response.message = "All dishes of this quantity successfully deleted from order"
-            return res.json(response)
+        const idValidator = ajv.compile(idRemoveValidator)
+        const valid = idValidator(item)
+        if (valid) {
+            try {
+                const removedOrder = await OrderService.removeOrderItem(db, item)
+                if (removedOrder.modifiedCount === 1) {
+                    let response  = JSONResponseService.generateSuccessResponse()
+                    response.message = "Item removed"
+                    response.data = removedOrder.ops
+                    return res.json(response)
+                }
+            } catch (e) {
+                let response = JSONResponseService.generateFailureResponse()
+                response.message = "An error has occurred"
+                return res.json(response)
+            }
         }
         let response = JSONResponseService.generateFailureResponse()
-        response.message = "Item was not deleted from order"
+        response.message = "Validator failed"
         return res.json(response)
-    } )
+    })
 }
 
 
+module.exports.editOrderItemQuantity = editOrderItemQuantity
 module.exports.createNewOrder = createNewOrder
-module.exports.addToOrder = addToOrder
+module.exports.getOrderDetails = getOrderDetails
 module.exports.removeOrderItem = removeOrderItem
+module.exports.submitFinalOrder = submitFinalOrder
+module.exports.addToOrder = addToOrder
+module.exports.getDishPriceById = getDishPriceById
